Refactor the whole immutable API to feel more like the mutable API
Instead of defining what type to reach from right to left the new api defines it from left to right.
This means that the final .to()-call on any immutable construction step has to conform to the current type, not the other way round.
Many of the previous problems have been fixed and we don't need map vs mapTo anymore, phew !